+++ tests/dummy/app/markdown/articles/one/details.md
@@ -1,0 +1,1 @@
+# Articles - one - details
