+++ tests/dummy/app/markdown/articles/one/index.md
@@ -1,0 +1,1 @@
+# Articles - one - index
