+++ tests/dummy/app/markdown/articles/three.md
@@ -1,0 +1,1 @@
+# Three
