+++ tests/dummy/app/markdown/articles/two/index.md
@@ -1,0 +1,1 @@
+# Articles - two - index
